Polish rulesheet and score scanner flows

+++ Pinball App Android/app/src/main/java/com/pillyliu/pinprofandroid/practice/ScoreScannerModels.kt
@@ -53,7 +53,7 @@ internal data class ScoreOcrObservation(
 
 internal data class ScoreScannerCandidate(
     val rawText: String,
-    val normalizedScore: Int,
+    val normalizedScore: Long,
     val formattedScore: String,
     val confidence: Float,
     val boundingBox: RectF,
@@ -67,7 +67,7 @@ internal data class ScoreScannerAnalysis(
 )
 
 internal data class ScoreScannerLockedReading(
-    val score: Int,
+    val score: Long,
     val formattedScore: String,
     val rawText: String,
     val confidence: Float,
